Add first version of kindelia place
Not very efficient right now. Cloning whole state on each read, which I don't think is necessary.

This file is too big for a starting block. To make testing easier, I just increased the allowed space of the blockchain. You either do the same or break it into smaller blocks.

I've included some auxiliary functions, but they would probably already be on the blockchain since they're very basic stuff.

This program will break if you pass in something that is not a number for any of the arguments. But for numbers, it does bound-checking.

A lot of thing are hardcoded for the 30x30 size. Could be more generic, but mana cost would go up.
I thought of doing one ctr with arity 30, but it'd be too ugly. Could be a good solution for scaling up dinamically (canvas of multiple 30x30 chunks).